app/client/requirements.txt
numpy
joblib
pillow
albumentations
fastapi[standard]
tensorflow
tensorflow-serving-api
py-grpc-prometheus